Fix deserialization error by changing positions field type from int to double
- Changed Positions property type from int to double in FinancialData model
- Updated CalculateProfitLossValue method signature to accept double positions
- Updated CalculateProfitLossPercentage method signature to accept double positions
- Fixes error: "The JSON value could not be converted to System.Int32. Path: $[4].positions"

The Infragistics API returns positions as a decimal number (double) rather than an integer, which was causing the deserialization failure at array index 4.

+++ src/samples/FinanceGrid.Library/Models/FinancialData.cs
@@ -14,7 +14,7 @@ public class FinancialData
     public PriceValue? Value { get; set; }
     
     [JsonPropertyName("positions")]
-    public int Positions { get; set; }
+    public double Positions { get; set; }
     
     [JsonPropertyName("profitLossValue")]
     public double ProfitLossValue { get; set; }
